+++ sync-for-commerce/src/codatsynccommerce/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.1'
-    sdk_version: str = '2.4.0'
-    gen_version: str = '2.142.2'
-    user_agent: str = 'speakeasy-sdk/python 2.4.0 2.142.2 1.1 codat-sync-for-commerce'
+    sdk_version: str = '2.4.1'
+    gen_version: str = '2.143.2'
+    user_agent: str = 'speakeasy-sdk/python 2.4.1 2.143.2 1.1 codat-sync-for-commerce'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: